refactor(cli): extract save/restore into commands/state_commands.py
Move the context save (currently registered as `dopemux backup` due to a
pre-existing decorator mismatch) and `dopemux restore` out of cli.py into
a dedicated state_commands module. Registration preserves the current
broken mapping so no behavior changes as part of extraction.

Update save/restore tests to patch ContextManager, Path.cwd, and
Path.exists at the new import site
(dopemux.commands.state_commands.*) and update the AST regression
guard for `restore` to search state_commands.py instead of cli.py.
